+++ scripts/macros/2024/mechanics/masteries.js
@@ -158,9 +158,9 @@ async function vex({workflow}) {
 }
 async function RollComplete(workflow) {
     if (!workflow.targets.size || !workflow.item || !workflow.activity || !workflow.actor || !workflow.token) return;
-    if (!constants.weaponAttacks.includes(workflow.activity.actionType)) return;
     let effects = actorUtils.getEffects(workflow.actor).filter(i => i.flags['chris-premades']?.mastery?.vex?.target === workflow.targets.first().id);
     if (effects.length) await genericUtils.deleteEmbeddedDocuments(workflow.actor, 'ActiveEffect', effects.map(i => i.id));
+    if (!constants.weaponAttacks.includes(workflow.activity.actionType)) return;
     let masterySetting = genericUtils.getCPRSetting('weaponMastery');
     let mastery;
     if (workflow.actor.type === 'npc') {