+++ scripts/applications/troubleshooter.js
@@ -217,6 +217,7 @@ export async function run() {
         addLine('GM Auto Roll Damage: ' + midiSettings.gmAutoDamage);
         addLine('Confirm Rolls: ' + midiSettings.confirmAttackDamage);
         addLine('Wait For Damage Application: ', midiSettings.waitForDamageApplication);
+        addLine('Auto Complete Workflow: ' + midiSettings.autoCompleteWorkflow);
     }
     try {
         if (game.modules.get('levelsautocover')?.active) {
@@ -461,6 +462,11 @@ async function startup() {
             addMidiMessage();
             content += '<li>' + genericUtils.translate('CHRISPREMADES.Troubleshooter.WaitForDamageApplication') + '</li>';
         }
+        if (!midiSettings.autoCompleteWorkflow) {
+            doSettingsMessage = true;
+            addMidiMessage();
+            content += '<li>' + genericUtils.translate('CHRISPREMADES.Troubleshooter.AutoCompleteWorkflow') + '</li>';
+        }
         if (midiAdded) content += '</ul>';
         if (game.modules.get('monks-wall-enhancement')?.active) {
             if (game.settings.get('monks-wall-enhancement', 'allow-one-way-doors')) {
@@ -526,7 +532,6 @@ async function startup() {
             content += genericUtils.translate('CHRISPREMADES.Troubleshooter.ModulesMessage') + '<ul>';
             let isFirst = true;
             incomptablemodulesFound.forEach(id => {
-                console.log(id);
                 if (!isFirst) content += '<br>';
                 content += '<li>' + game.modules.get(id).title + '</li>';
                 isFirst = false;
@@ -582,11 +587,12 @@ async function fixSettings(message) {
         autoCEEffects: 'itempri',
         allowActorUseMacro: true,
         allowUseMacro: true,
-        autoRollDamage: 'onHit',
-        gmAutoDamage: 'onHit',
         confirmAttackDamage: 'none',
-        waitForDamageApplication: true
+        waitForDamageApplication: true,
+        autoCompleteWorkflow: true
     };
+    if (!['onHit', 'always'].includes(midiSettings.autoRollDamage)) newMidiSettings.autoRollDamage = 'onHit';
+    if (!['onHit', 'always'].includes(midiSettings.gmAutoDamage)) newMidiSettings.gmAutoDamage = 'onHit';
     Object.entries(newMidiSettings).forEach(([key, value]) => genericUtils.setProperty(midiSettings, key, value));
     await game.settings.set('midi-qol', 'ConfigSettings', midiSettings);
     if (game.modules.get('monks-wall-enhancement')?.active) {