+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 ## New Generic Monster Features:
 - Failed by Amount
 - Choose Damage Type
+## Update Notes:
+- Updated Troubleshooter
+- Added module and module setting checks.
   
 # 1.3.38 Release:
 ## New Generic Monster Features: